Add user interface (#8)
* Changed the API for the breadcrumbs.

The functions now compose (in the common case) like:

(add-breadcrumb! (make-breadcrumb! message category))

There is a beginning of support for passing an explicit context instead
of relying on the thread-local save mechanism, but that part is not
finished with this commit. The above case is working however.

* Add support for explicit context.

Users can now pass an explicit context as the first parameter to the
public functions. This allows to bypass the use of thread-local storage.

The http client instance previously passed as first argument must now be
supplied as a :http entry in the context map instead.


* Added support for sentry's user interface.

It is now possible to add user information using the newly introduced
add-user and make-user functions.