Merge :params into the env dict.
As reported in https://github.com/exoscale/raven/issues/40 Sentry does
not seem to display nested dicts nicely in its http interface.

Since our helper here is advertising to work with ring requests, we
should handle the case of the user using ring.middleware.params
gracefully and make it look as nice as we can. Therefore, we now merge
that dict into the "env" sentry dict, so as to have one nice entry per
item.